incorporated some of the comments of the reviewers
UNINCORPORATED COMMENTS:

Reviewer 1

>  I would also appreciate if the authors would cite a paper about contracts
>  in Scheme instead of referring to the PLT manual.
No idea which one.

> On the end of page 4, the remark about intermediate forms sounds a lot like
> typed intermediate forms so I wondered how a intermediate form with
> contracts compares to a typed intermediate form.
Not knowledgeable about typed intermediate forms.

> "(define bst (bounded-bst ...))": why is the contract not called "bst?"?
I think it is obvious enough that bst? is already bound to the predicate. We
have referred to it as the predicate in the text.

Reviewer 2

> - Section 5.1: The left figure of Fig 7 appears to be linear.
>   Shouldn't it be exponential?  Is there any explanation for that?
The graph's y-axis is labeled as "factor slowdown", so a linear shape for that
means exponential slowdown of run time. Is it possible to actually just plot
the slowdown for a more obvious graph?

> - Page 16, line 5, "the attributes themselves are not used in the
>   semantics of the stronger relation": I did not understand what this
>   sentence says.  It appears to me that stronger relation of between/c
>   does use the values of the bounds.  Accordingly, I think you need a
>   kind of symbolic evaluation to determine the stronger relation once
>   you introduced the unknown value "?".  For example, (between/c 2 ?)
>   is stronger than (between/c 2 (+ ? 1)) only when you can prove
>   symbolically that (< ? (+ ? 1)) for any "?".
The reason the stronger relations works is that these attributes necessarily
depend on dependent contracts - so for stronger only the closure pointer
comparison is used. Is that right?

Reviewer 3

> Section 3 mentions lots of data structure invariants that have been
> explored. The question is how? Did you write down each invariant as a
> contract (where?) or did you even implement all of these data structures?
Unsure how to change.

> Implementation: Why do you copy nodes into the node-wrapper? Why not after
> checking the contract replace the node-wrapper by the original node with the
> pointers redirected to the new node-wrappers of the children? This look like
> potentially a lot of copying and the wrappers waste a lot of space for
> values only filled in later.
Unless we change the implementation, leaving this unincorporated.

> Section 7: Which ideas do you consider carrying back to your setting to
> improve error messages?
What do you think, Robby?

> Most English style guides recommend using "since" only for its original
> temporal meaning and otherwise use "because".
I have no problems with this. Natural language has lexical ambiguity all over
the place, across the board.

> Can this extension of PLT Scheme be downloaded somewhere?
I am not sure which version started carrying the lazy data structure code.
Once I find out I will add a sentence in the footnote that states "We use the
PLT Scheme contract library throughout."

Reviewer 4

> - it would be nice if you could polish a bit more the paragraph on
>   the relation relation between your approach and that of Chitil,
>   Huch, et al.
Have not yet read the paper.